requirements.txt
filelock==3.19.1
fsspec>=2024.1.1
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pillow==11.3.0
sympy==1.14.0

# PyTorch stack
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

typing_extensions==4.15.0
